+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix typing in LLM (#568)
+- Fix parsing functions with no arguments in convert_function_to_function_schema (#568)
 - Remove outdated quickstart ref from docstring (#565)
 - Add support for async component factories (#608)
 - Add function calling for LLMs (#533)